% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdeconv.R
\name{cellscatterplots}
\alias{cellscatterplots}
\title{Draw scatter plots to compare two sets of cell content values}
\usage{
cellscatterplots(
  dat1,
  dat2,
  title = NULL,
  colorful = TRUE,
  corcmp = TRUE,
  usePCC = FALSE,
  xtitle = "RNA cell contents",
  ytitle = "Methylation cell contents",
  annotextsize = 6,
  textsize = 13,
  titlesize = 15,
  face = "bold",
  dotsize = 1
)
}
\arguments{
\item{dat1}{A data frame or matrix recording the sample cell contents in the
first dataset. Each column is a cell type and each row is a sample. The
column names are the cell type names and the row names are the sample IDs.}

\item{dat2}{A data frame or matrix recording the cell contents for the same
samples in the second dataset. Each column is a cell type and each row is
a sample. The column names are the cell type names and the row names are
the sample IDs.}

\item{title}{A string to define the prefixes of the plot titles, can also be
set as NULL.}

\item{colorful}{A logical value indicating whether the dots in the scatter
plots should be rainbow colored to indicate their density or not. Default
is TRUE. This parameter only works when the dot number in the plot is more
than 50, otherwise, all the dots will be colored as blue.}

\item{corcmp}{A logical value indicating whether a scatter plot comparing
the cell-cell correlation in the 2 datasets should be drawn or not. The
default value is TRUE.}

\item{usePCC}{If this value is TRUE, the Pearson correlation coefficient
between the cell contents will be labeled in the plot. If FALSE, the R
square value will be labeled instead. Default is FALSE.}

\item{xtitle}{The title of the x-axes of the scatter plots. Default is "RNA
cell contents".}

\item{ytitle}{Title of the y-axes. Default is "Methylation cell contents".}

\item{annotextsize}{A number used to define the font size of the annotation
labels in the scatter plot. Default is 6.}

\item{textsize}{A number to define the base font size for the plots, except
the plot title size, axis title size, and the annotation label size, which
are defined by code{titlesize} and \code{annotextsize}. The default value
is 13.}

\item{titlesize}{A number to define the size for the plot and axis titles,
and the default value is 15.}

\item{face}{A string to define whether the text in the plot should be bold
or plain. Default is "bold".}

\item{dotsize}{A number to define the dot size in the scatter plots. Default
is 1.}
}
\value{
Generate a scatter plot to compare the cell contents for the same
samples recorded by two datasets. For each cell type, a facet will be made
for its plot. If the parameter \code{corcmp} is set as TRUE, an additional
scatter plot comparing the cell-cell correlation between the two datasets
will also be drawn.
}
\description{
Draw scatter plots to compare two sets of cell content values and compare
two sets of cell-cell correlation values. For each cell type, its scatter
plot accounts for one facet of all the plots.
}
\examples{
scRNA <- system.file('extdata', 'scRNAseqdat.rds', package = 'scDeconv')
scRNA <- readRDS(scRNA)

pRNA <- system.file('extdata', 'pairedRNAdat.rds', package = 'scDeconv')
pRNA <- readRDS(pRNA)

pDNAm <- system.file('extdata', 'pairedDNAmdat.rds', package = 'scDeconv')
pDNAm <- readRDS(pDNAm)

pseudobulk <- system.file('extdata', 'pseudobulk.rds', package = 'scDeconv')
pseudobulk <- readRDS(pseudobulk)

refres <- scRef(Seuratobj = scRNA,  
               targetcelltypes = c('EVT', 'FB', 'HB', 'VCT'),  
               celltypecolname = 'annotation',  
               pseudobulkdat = pseudobulk, 
               targetdat = pRNA, 
               targetlogged = TRUE)

dnamres <- epDeconv(rnaref = refres$ref, 
                   rnamat = refres$targetnolog, 
                   rnamatlogged = FALSE, 
                   methylmat = pDNAm, 
                   learnernum = 10, 
                   resscale = TRUE)

cellscatterplots(dat1 = dnamres$rnacellconts, 
                dat2 = dnamres$methylcellconts, 
                xtitle = 'RNA cell contents',
                ytitle = 'Methylation cell contents')
}
