% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction.R
\name{enrichwrapper}
\alias{enrichwrapper}
\title{Annotate the function of the cell type specific differential CpG sites}
\usage{
enrichwrapper(
  sigprobes,
  celltype,
  pairedRNA = NULL,
  pairedmethyl = NULL,
  abscut = 0.6,
  generegions = c("TSS200"),
  platform = "450K",
  uniquegenes = TRUE,
  dbs = c("GO_Biological_Process_2018", "BioPlanet_2019", "Reactome_2016"),
  write = FALSE
)
}
\arguments{
\item{sigprobes}{The list result returned by \code{celldiff} with the inter-
group differential probe selection results for each cell type.}

\item{celltype}{The cell type whose differential methylation probes needs to
be annotated.}

\item{pairedRNA}{The RNA part of the paired RNA-DNA methylation dataset. If
both this parameter and \code{pairedmethyl} are not NULL, a correlation-
based method will be used to find the genes whose RNA expression level in
the RNA data significantly correlated with any of the cell type specific
differential probes in the paired methylation data, and then the enriched
function of these selected genes will be analyzed and each of them will
also has its individual function annotated. The default value is NULL, and
in this case, the correlation-base method will not be used, and instead,
the function annotation and enrichment analysis will be directly conducted
on the cell type specific hyper and hypomethylated DNA methylation probes
in \code{sigprobes}, with the probes mapped to corresponding genes first.}

\item{pairedmethyl}{The methylation part of the paired RNA-DNA methylation
dataset. If both this parameter and \code{pairedRNA} are not NULL, the
correlation-based method will be used to find the genes with an expression
level significantly correlated to the differential methylation probes and
perform function analysis on them, while if it is NULL, the analysis will
be conducted directly on the genes corresponding to the cell type specific
differential methylation probes in \code{sigprobes}. Default is NULL.}

\item{abscut}{When the correlation method is used to select the genes with a
significant correlation to the cell type specific differential probes, the
ones with an absolute Pearson correlation coefficient value greater than
this parameter value will be selected and used for function analysis. The
default value is 0.6.}

\item{generegions}{When the function analysis is directly performed on the
genes mapped from the differential methylated probes, if a probe located
in the regions defined by this parameter, its corresponding gene will be
used for the function analysis. Default is "TSS200", so that only probes
within TSS200 regions will be used for gene mapping, and it can also be a
vector, such as \code{c("TSS200", "TSS1500", "1stExon")}, so that probes
within any of these 3 regions will be used for gene mapping.}

\item{platform}{A string indicating the platform of the differential probes
recorded in \code{sigprobes}. Default is "450K", can also be "EPIC".}

\item{uniquegenes}{A logical value and if this parameter is set as TRUE,
after getting the genes for function analysis from the cell type indicated
by \code{celltype}, these genes will be used to compare with the genes
obtained similarly from other cell types recorded in \code{sigprobes}, and
only the ones uniquely contained in the target cell type will be used for
the downstream analysis. If this parameter is NULL, this filter step will
not be performed. Default is TRUE.}

\item{dbs}{The databases for gene function enrichment analysis. Default is
\code{c("GO_Biological_Process_2018", "BioPlanet_2019", "Reactome_2016")}.}

\item{write}{A logical value indicating whether the gene function results
need to be written into txt files in the working directory. The default
value is FALSE.}
}
\value{
A list with four slots recording the gene annotation and function
enrichment results for the enhanced and inhibited genes (when correlation-
based method is used) or for the genes mapped from the hypomethylated and
hypermethylated probes (when correlation-based method is not used), and if
\code{write} is TRUE, txt files will be made to save these results.
}
\description{
Annotate the function of the cell type specific differential DNA methylation
probes called from \code{celldiff}.
}
