% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epdeconv.R
\name{methylpredict}
\alias{methylpredict}
\title{Predict cell contents for new methylation data using an ensemble model}
\usage{
methylpredict(
  model = NULL,
  normweights = NULL,
  modellist = NULL,
  targetmethyldat,
  resscale = FALSE,
  adjustminus = FALSE
)
}
\arguments{
\item{model}{Result returned from the function \code{epDeconv}. It contains
the deconvoltion ensemble model. Default value is NULL, and in this case,
the other two parameters \code{normweights} and \code{modellist} will be
used instead to finish the prediction, if they are not NULL.}

\item{modellist}{The slot "modellist" of the result of \code{epDeconv}, and
it provides the base learners of the deconvolution ensemble model. If the
parameter \code{model} is NULL, it will be used to restore the model, and
finish the prediction. Default value is NULL.}

\item{targetmethyldat}{The target cell mixture methylation data need to be
deconvolved. Should be a matrix with each column representing one sample
and each row for one feature. Row names are feature names and column names
are sample IDs. It is recommended to adjust the batch difference between
this dataset and the paired methylation data used by \code{epDeconv} with
\code{ComBat} in advance, and using the paired data as the reference batch
when adjusting, so that the cell deconvolution model trained can be used
on these target data with the influence from batch difference minimized.}

\item{resscale}{For each sample, whether its cell contents result should be
scaled so that the sum of different cell types is 1. Default is FALSE.}

\item{adjustminus}{In some extreme situations, the cell contents predicted
may be minus values and this parameter indicates whether these minus ones
should be changed to 0 in the final result. Default is FALSE.}

\item{normweight}{The slot "normweigth" of the result of \code{epDeconv}. It
provides the base learner weights of the deconvolution ensemble model. If
\code{model} is NULL, this parameter will be used to restore the model and
finish the prediction. Default value is NULL.}
}
\value{
A matrix recording the cell composition result for the samples.
}
\description{
Predict cell contents for new methylation data with the model trained from
the function \code{epDeconv}.
}
\examples{
scRNA <- system.file('extdata', 'scRNAseqdat.rds', package = 'scDeconv')
scRNA <- readRDS(scRNA)

pRNA <- system.file('extdata', 'pairedRNAdat.rds', package = 'scDeconv')
pRNA <- readRDS(pRNA)

pDNAm <- system.file('extdata', 'pairedDNAmdat.rds', package = 'scDeconv')
pDNAm <- readRDS(pDNAm)

externalDNAm <- system.file('extdata', 'externalDNAmdat.rds', package = 'scDeconv')
externalDNAm <- readRDS(externalDNAm)

pseudobulk <- system.file('extdata', 'pseudobulk.rds', package = 'scDeconv')
pseudobulk <- readRDS(pseudobulk)

refres <- scRef(Seuratobj = scRNA,  
               targetcelltypes = c('EVT', 'FB', 'HB', 'VCT'),  
               celltypecolname = 'annotation',  
               pseudobulkdat = pseudobulk, 
               targetdat = pRNA, 
               targetlogged = TRUE)

dnamres <- epDeconv(rnaref = refres$ref, 
                   rnamat = refres$targetnolog, 
                   rnamatlogged = FALSE, 
                   methylmat = pDNAm, 
                   learnernum = 10, 
                   resscale = TRUE)
                   
externalcellconts <- methylpredict(model = dnamres, 
                                  targetmethyldat = externalDNAm, 
                                  resscale = TRUE, 
                                  adjustminus = TRUE)
}
