% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction.R
\name{probetogene}
\alias{probetogene}
\title{Summarize the methylation beta values of probes to genes}
\usage{
probetogene(
  betadat,
  platform = "450K",
  range27k = 200,
  group450k850k = "TSS200",
  includemultimatch = FALSE
)
}
\arguments{
\item{betadat}{A matrix recording the beta values of methylation probes for
samples. Each column represents one sample and each row represents one
probe. The row names are the probe names while the column names should be
sample IDs.}

\item{platform}{The platform of the probes. Can be set as "27K", "450K", or
"EPIC". Default is "450K".}

\item{range27k}{A positive number or a vector with two positive numbers. If
the data is from 27K platform, this parameter is needed to define which
probes should be considered as related to a specific gene, and only the
ones with a distance to the TSS of a gene less than the maximum value and
greater than the minimum value of \code{range27k} will be considered as
related to the gene, and the beta values of these probes will be averaged
to get the gene beta value. If it is a single number, the probes with a
distance less than this number and greater than 0 will be attributed to a
gene. Default is 200.}

\item{group450k850k}{A vector or single string. If the data is based on 450K
or EPIC platform, this parameter is needed to define which probes could be
considered as related to a specific gene. Only the ones located in the
gene regions included in this parameter will be considered as belong to
the gene. The value of this parameter need to be selected from "TSS200",
"TSS1500", "1stExon", "5'UTR", '3'UTR", and "Body". If it is a vector,
such as \code{c("TSS200", "TSS1500", "1stExon")}, the probes within these
3 regions of a gene will be attributed to the gene and their beta values
will be averaged to get the gene beta value. Default value is the string
"TSS200".}

\item{includemultimatch}{Some probes can be attributed to multiple genes.
If this parameter is TRUE, these probes will be involved into the beta
value calculation for all their related genes. Otherwise, these probes
will be discarded, so that the beta values of all the genes are averaged
only from their uniquely related probes. Default is FALSE.}
}
\value{
A matrix recording the summarized gene beta values for samples.
}
\description{
Summarize the methylation beta values of probes to genes by averaging the
probes located closely to the TSS of a gene.
}
\examples{
pDNAm <- system.file('extdata', 'pairedDNAmdat.rds', package = 'scDeconv')
pDNAm <- readRDS(pDNAm)

pDNAmgenes <- probetogene(betadat = pDNAm, 
                         platform = "450K", 
                         group450k850k = "TSS200")
}
