% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdeconv.R
\name{refDeconv}
\alias{refDeconv}
\title{Deconvolve cell contents using reference from the same omics}
\usage{
refDeconv(
  ref,
  targetdat,
  targetlogged = FALSE,
  resscale = FALSE,
  plot = FALSE,
  pddat = NULL,
  threads = 1
)
}
\arguments{
\item{ref}{The reference matrix recording the signature of each cell type.
Each row represents one feature, and each column is one cell type. Each
entry should be a non-log transformed value, such as TPM value for RNA
data, and beta value for DNA methylation data. Column names are cell type
names and row names are feature names.}

\item{targetdat}{The target cell mixture data need to be deconvolved. Should
be a matrix with each column representing one sample and each row for one
feature. Row names are feature names and column names are sample IDs. If
the reference matrix is generated with the function \code{scRef}, and both
the scRNA-seq and the target cell mixture data were transferred to it, the
result reference matrix can be transferred to \code{ref} and the adjusted
target data returned by \code{scRef} can be transferred to this parameter.}

\item{targetlogged}{Whether the feature values in \code{targetdat} are log2
transformed values or not.}

\item{resscale}{For each sample, whether its cell contents result should be
scaled so that the sum of different cell types is 1. Default is FALSE.}

\item{plot}{Whether generate a box plot and heatmaps for the cell contents
deconvolved. Default is FALSE.}

\item{pddat}{If set \code{plot} as TRUE, this parameter can be used to show
the sample group information, so that the box plot generated will also
compare the group difference for each cell type, and heatmaps with this
comparison will also be generated. It should be a data frame recording the
sample groups, and must include 2 columns. One is named as "sampleid",
recording the sample IDs same as the column names of \code{targetdat}, the
other is "Samplegroup", recording the sample group to which each sample
belongs. It can also be NULL, meaning all the samples are from the same
group, and in this case, only a box plot showing the cell content for each
cell type will be made when \code{plot} is set as TRUE.}

\item{threads}{Number of threads need to be used to do the computation. Its
default value is 1.}
}
\value{
A matrix recording the cell composition result for the samples.
}
\description{
Deconvolve cell contents using reference from the same omics.
}
