% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction.R
\name{celldiff}
\alias{celldiff}
\title{Find cell type specific differential features from bulk data}
\usage{
celldiff(
  cellconts,
  vardat,
  responsedat,
  padjcutoff = NULL,
  pcutoff = NULL,
  gradientcutoff = NULL,
  threads = 1,
  int = TRUE
)
}
\arguments{
\item{cellconts}{A matrix recording the cell contents of the samples. Each
row is a sample and each column is a cell type. The row names are sample
IDs and the column names are cell type names. The deconvolution result
from \code{refDeconv} and \code{methylpredict} can be directly used here.}

\item{vardat}{A data frame recording the sample group information, and must
include 2 columns. One is named as "sampleid", recording the sample IDs
same as the row names of \code{cellconts}, the other is "Samplegroup",
recording the sample group to which each sample belongs. If there are any
additional columns in this data frame, they will be deemed as confounding
factors when selecting the cell type specific inter-group differential
features using a linear regression model.}

\item{responsedat}{A matrix reconding the feature values of the samples, and
each row is a feature and each column is a sample. The column names are
sample IDs and the row names are feature names. The cell type specific
differential features will be selected from the features in this matrix.}

\item{padjcutoff}{The adjusted p-value cutoff to select the significantly
differential features. Default is NULL, and in this case, the original
p-value will be used to select the differential features instead and its
cutoff is defined by the parameter \code{pcutoff}, but if \code{pcutoff}
is also NULL, the criterion of adjusted p-value < 0.05 will be used to
find the differential features.}

\item{pcutoff}{If \code{padjcutoff} is set as NULL, this parameter will be
used to set a cutoff on the original p-value to select the significantly
differential features. It default value is also NULL, and in the case that
both \code{padjcutoff} and \code{pcutoff} are NULL, \code{padjcutoff} will
be set as 0.05 automatically and used to find the differential features.}

\item{gradientcutoff}{The cutoff on the gradient (i.e. the coefficient of
the interaction term in the regression model between a specific cell type
content and the sample group, which reflects the group difference on the
partial derivative of the feature to the cell content) to find the cell
type specific differential features. The default is NULL, and it will be
deemed as 0. The cell type specific differential features between sample
groups are selected using a regression model containing an interaction
term between the sample group and the specific cell type content.}

\item{threads}{Number of threads need to be used to do the computation. Its
default value is 1.}

\item{int}{A logical value indicating whether the cell contents need to be
transformed to fit an inverse normal distribution before the differential
feature regression model construction, so that the collinearity caused by
the interaction term in the model can be reduced. Default is TRUE.}
}
\value{
A list with various slots recording the inter-group differential
feature selection results for each cell type. Each slot contains a matrix
and the feature names, p-values, adjusted p-values, etc, are included. In
addition, volcano plots for the differential features for each cell type
will also be generated.
}
\description{
Find cell type specific inter-group differential features from bulk data and
cell content information using a regression model containing an interaction
term between the sample group and the specific cell type content.
}
\examples{
scRNA <- system.file('extdata', 'scRNAseqdat.rds', package = 'scDeconv')
scRNA <- readRDS(scRNA)

pRNA <- system.file('extdata', 'pairedRNAdat.rds', package = 'scDeconv')
pRNA <- readRDS(pRNA)

pDNAm <- system.file('extdata', 'pairedDNAmdat.rds', package = 'scDeconv')
pDNAm <- readRDS(pDNAm)

externalDNAm <- system.file('extdata', 'externalDNAmdat.rds', package = 'scDeconv')
externalDNAm <- readRDS(externalDNAm)

DNAmpd <- system.file('extdata', 'DNAmpd.rds', package = 'scDeconv')
DNAmpd <- readRDS(DNAmpd)

pseudobulk <- system.file('extdata', 'pseudobulk.rds', package = 'scDeconv')
pseudobulk <- readRDS(pseudobulk)

refres <- scRef(Seuratobj = scRNA,  
               targetcelltypes = c('EVT', 'FB', 'HB', 'VCT'),  
               celltypecolname = 'annotation',  
               pseudobulkdat = pseudobulk, 
               targetdat = pRNA, 
               targetlogged = TRUE)

dnamres <- epDeconv(rnaref = refres$ref, 
                   rnamat = refres$targetnolog, 
                   rnamatlogged = FALSE, 
                   methylmat = pDNAm, 
                   learnernum = 10, 
                   resscale = TRUE, 
                   targetmethyldat = externalDNAm)
                   
totalcellconts <- rbind(dnamres$methylcellconts, 
                       dnamres$methyltargetcellcounts)

totalvardat <- DNAmpd[c("sampleid", "Samplegroup", "Gestwk")]

totalresponse <- cbind(pDNAm, externalDNAm)

celldiffprobes <- celldiff(cellconts = totalcellconts,
                          vardat = totalvardat, 
                          responsedat = totalresponse, 
                          pcutoff = 0.01, 
                          gradientcutoff = 0)
}
